fix(webrtc): wait for signaling transport when subscribing
The subscription routines (camera, screenshare) are not waiting for the
signaling transport to be properly established before being run. That
causes a client crash.

This commit tracks the signaling transport state in the subscription
routines.

The ideal solution would be to queue up RPC calls in the media managers
themselves and flush them when the socket is open. That will be
addressed soon.