fix(audio): handle permission issues
There's no handler in place for lack-of-permission issues when joining
audio.

This commit introduces an alert explaining what happened and offering
the same course of actions as the ones provided by cde0f749808.

It also moves the join audio procedure to a thunk to makes things easier
and allow us to decouple AudioManager from the store in the near future.

An issue with joinAudio on login where it wouldn't abide to the room's
lock settings is also fixed.