fix: tie log in states to token validation AND collection subscriptions
client.sessionState.loggedIn/loggingIn currently just tracks the
token validation routine, which means it can be flagged as logged
in before collections are populated. It should be a reliable global
state to check whether the client is ready to do anything and it isn't
right now.

This commits addresses the issue via a few changes:
  - Extend transactions to subscribe requests as well so we can track
    them E2E
  - Extend slices that represent remote collections with a ready flag +
    readyStateChanged reducer to flag them as ready/subscribed
  - Use those ready states for some basic collections alongside the auth
    token validation state to flag the client as loggedIn.