fix(audio): route audio through speaker by default
Audio comes out of the ear phone by default which is not
desirable. Needs to come from the speaker by default and then switch two
headset (jack or bluetooth) when plugged.

This commit addresses that through a rather unorthodox use of expo-av's
setAudioMode#allowsRecordingIOS and
setAudioMode#playThroughEarpieceAndroid to guarantee audio comes out of
the speaker by default, but switch to headsets when they're plugged in.
For the headset plug in/out detection, I included
https://www.npmjs.com/package/react-native-headphone-detection - which
is not that well maintained, but seems to be doing its job properly.