feat(feedback): adds 3 gathering screens
This commit adds 3 feedback screens:
  - Star rating
  - Problem
  - Email gathering
The feedback has 2 modes, which are controlled by the 'feedback.enabled' and
'feedback.customFeedback.enabled' flags in the settings.json file.
Also, this commit moves all the logic of the end reason message, from the end
session screen to a hook so it can be used in the feedback screens.