fix(audio): audio incorrectly re-joins on spurious re-render
The client is re-rendering some components due to an impure prop being
updated. The offender prop in question is not known right now, but due
to weak checks in the LiveKit component that define when audio should
be automatically activated, it triggered audio re-joins in scenarios
like changing the room's recording state.

Prevent unexpected audio re-joins by:
  - Taking audio connection states into account in the join use effect
  - Improving how trailing media manager init requests are handled

Additionally, add an uncommited patch that improves LiveKit's handling
of inactive media stream.