refactor(audio): trim useAudioJoin props, use it as dep in callers
Trim the useAudioJoin hook dependencies so that it isn't re-created
unnecessarily - i.e.: when the meeting object change any unused prop.

Additionally, make sure the hook's users call it as a dependency
whenever applicable (either in a useEffect or via a useCallback), so
that we guarantee that no expected behavior is present.