fix: broken lock settings caused WebRTC media to fail
The way lock settings data is fetched was not migrated to graphql/3.0.
This causes it to throw an exception as it is relying on old Redux data
structures that are not fully available anymore.
That exception causes media initialization to fail whenever a meeting has
lock settings active AND the user is locked.

Migrate lock settings data fetch to GraphQL, restoring lock settings
functionality and fixing the media issues mentioned above.